refactor(risk): move pure risk types to domain (D1)
RiskContext, RiskDecision and KillSwitchSnapshot are pure value types and now live in domain/risk, alongside the other domain models. RiskGate (logic) and the @Component KillSwitchState (stateful) stay in application/risk. Pure package/import churn, no behavior change; all references updated.
