+++ src/main/kotlin/ru/driics/aitrade/infra/risk/FileKillSwitchStore.kt
@@ -2,8 +2,8 @@ package ru.driics.aitrade.infra.risk
 
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import com.fasterxml.jackson.module.kotlin.readValue
-import ru.driics.aitrade.application.risk.KillSwitchSnapshot
 import ru.driics.aitrade.application.risk.KillSwitchStore
+import ru.driics.aitrade.domain.risk.KillSwitchSnapshot
 import java.nio.file.AtomicMoveNotSupportedException
 import java.nio.file.Files
 import java.nio.file.Path
