+++ src/test/kotlin/ru/driics/aitrade/application/risk/KillSwitchStatePersistenceTest.kt
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.Assertions.assertFalse
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.io.TempDir
+import ru.driics.aitrade.domain.risk.KillSwitchSnapshot
 import ru.driics.aitrade.infra.risk.FileKillSwitchStore
 import java.nio.file.Files
 import java.nio.file.Path
