+++ src/test/kotlin/ru/driics/aitrade/application/usecase/ExecuteAiDecisionsUseCaseTest.kt
@@ -10,10 +10,10 @@ import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
 import ru.driics.aitrade.application.ai.ConfidenceCalibrator
-import ru.driics.aitrade.application.risk.KillSwitchSnapshot
-import ru.driics.aitrade.application.risk.RiskContext
-import ru.driics.aitrade.application.risk.RiskDecision
 import ru.driics.aitrade.application.risk.RiskGate
+import ru.driics.aitrade.domain.risk.KillSwitchSnapshot
+import ru.driics.aitrade.domain.risk.RiskContext
+import ru.driics.aitrade.domain.risk.RiskDecision
 import ru.driics.aitrade.config.RiskGateProperties
 import ru.driics.aitrade.config.TradingProperties
 import ru.driics.aitrade.domain.model.AIAction
